+++ src/output.rs
@@ -1,45 +1,52 @@
 // Copyright 2021 Martin Pool
 
-//! A `target/mutants` directory holding logs and other output.
+//! A `mutants.out` directory holding logs and other output.
 //!
 //! *CAUTION:* This currently doesn't interact with Cargo locking, and if two `cargo-mutants`
 //! processes access the same directory they'll tread on each other...
 
 use std::fs::File;
 use std::io::{Read, Seek};
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::{fs, io};
 
 use anyhow::{Context, Result};
 
-use crate::source::SourceTree;
+const OUTDIR_NAME: &str = "mutants.out";
+const ROTATED_NAME: &str = "mutants.out.old";
 
+/// A `mutants.out` directory holding logs and other output information.
 #[derive(Debug)]
 pub struct OutputDir {
-    pub path: PathBuf,
-    pub log_dir: PathBuf,
+    path: PathBuf,
+    log_dir: PathBuf,
 }
 
 impl OutputDir {
-    pub fn new(tree: &SourceTree) -> Result<OutputDir> {
-        let path: PathBuf = tree.root().join("target").join("mutants");
-        fs::create_dir_all(&path)
-            .with_context(|| format!("create output directory {:?}", &path))?;
+    /// Create a new `mutants.out` output directory, within the given directory.
+    ///
+    /// If the directory already exists, it's rotated to `mutants.out.old`. If that directory
+    /// exists, it's deleted.
+    pub fn new<P: AsRef<Path>>(in_dir: P) -> Result<OutputDir> {
+        let path: PathBuf = in_dir.as_ref().join(OUTDIR_NAME);
+        if path.exists() {
+            let rotated = in_dir.as_ref().join(ROTATED_NAME);
+            if rotated.exists() {
+                fs::remove_dir_all(&rotated).with_context(|| format!("remove {:?}", &rotated))?;
+            }
+            fs::rename(&path, &rotated)
+                .with_context(|| format!("move {:?} to {:?}", &path, &rotated))?;
+        }
+        fs::create_dir(&path).with_context(|| format!("create output directory {:?}", &path))?;
         let log_dir = path.join("log");
-        fs::create_dir_all(&log_dir)
-            .with_context(|| format!("create log directory {:?}", &log_dir))?;
+        fs::create_dir(&log_dir).with_context(|| format!("create log directory {:?}", &log_dir))?;
         Ok(OutputDir { path, log_dir })
     }
 
-    pub fn delete_logs(&self) -> Result<()> {
-        for entry in self.log_dir.read_dir()? {
-            let path = entry?.path();
-            fs::remove_file(&path).with_context(|| format!("delete log file {:?}", &path))?;
-        }
-        Ok(())
-    }
-
+    /// Create a new log for a given scenario.
     pub fn create_log(&self, scenario_name: &str) -> Result<TestLog> {
+        // TODO: Maybe remember what files have already been created to avoid this loop, although
+        // realistically it seems unlikely to be hit often...
         let basename = clean_filename(scenario_name);
         for i in 0..1000 {
             let t = if i == 0 {
@@ -63,6 +70,12 @@ impl OutputDir {
             "couldn't create any test log in {:?} for {:?}",
             self, scenario_name,
         );
+    }
+
+    #[allow(dead_code)]
+    /// Return the path of the `mutants.out` directory.
+    pub fn path(&self) -> &Path {
+        &self.path
     }
 }
 
@@ -99,6 +112,7 @@ mod test {
     use tempfile::TempDir;
 
     use super::*;
+    use crate::source::SourceTree;
 
     fn minimal_source_tree() -> TempDir {
         let tmp = tempfile::tempdir().unwrap();
@@ -106,46 +120,67 @@ mod test {
         tmp
     }
 
-    #[test]
-    fn create() {
-        let tmp = minimal_source_tree();
-        let src_tree = SourceTree::new(tmp.path()).unwrap();
-        let _output_dir = OutputDir::new(&src_tree).unwrap();
-        let names = walkdir::WalkDir::new(tmp.path())
+    fn list_recursive(path: &Path) -> Vec<String> {
+        walkdir::WalkDir::new(path)
             .sort_by_file_name()
             .into_iter()
             .map(|entry| {
                 entry
                     .unwrap()
                     .path()
-                    .strip_prefix(tmp.path())
+                    .strip_prefix(path)
                     .unwrap()
                     .to_slash_lossy()
                     .to_string()
             })
-            .collect_vec();
-
-        assert_eq!(
-            names,
-            &[
-                "",
-                "Cargo.toml",
-                "target",
-                "target/mutants",
-                "target/mutants/log"
-            ]
-        );
+            .collect_vec()
     }
 
     #[test]
-    fn delete_existing_logs() {
+    fn create() {
         let tmp = minimal_source_tree();
         let src_tree = SourceTree::new(tmp.path()).unwrap();
-        let output_dir = OutputDir::new(&src_tree).unwrap();
-        let log_file_path = output_dir.log_dir.join("something.log");
-        fs::write(&log_file_path, b"stuff\n").unwrap();
-        assert!(log_file_path.is_file());
-        output_dir.delete_logs().unwrap();
-        assert!(!log_file_path.is_file());
+        let output_dir = OutputDir::new(src_tree.root()).unwrap();
+        assert_eq!(
+            list_recursive(tmp.path()),
+            &["", "Cargo.toml", "mutants.out", "mutants.out/log",]
+        );
+        assert_eq!(output_dir.path(), tmp.path().join("mutants.out"));
+        assert_eq!(output_dir.log_dir, tmp.path().join("mutants.out/log"));
+    }
+
+    #[test]
+    fn rotate() {
+        let temp_dir = tempfile::TempDir::new().unwrap();
+
+        // Create an initial output dir with one log.
+        let output_dir = OutputDir::new(&temp_dir).unwrap();
+        output_dir.create_log("one").unwrap();
+        assert!(temp_dir.path().join("mutants.out/log/one.log").is_file());
+
+        // The second time we create it in the same directory, the old one is moved away.
+        let output_dir = OutputDir::new(&temp_dir).unwrap();
+        output_dir.create_log("two").unwrap();
+        assert!(temp_dir
+            .path()
+            .join("mutants.out.old/log/one.log")
+            .is_file());
+        assert!(temp_dir.path().join("mutants.out/log/two.log").is_file());
+        assert!(!temp_dir.path().join("mutants.out/log/one.log").is_file());
+
+        // The third time (and later), the .old directory is removed.
+        let output_dir = OutputDir::new(&temp_dir).unwrap();
+        output_dir.create_log("three").unwrap();
+        assert!(temp_dir.path().join("mutants.out/log/three.log").is_file());
+        assert!(!temp_dir.path().join("mutants.out/log/two.log").is_file());
+        assert!(!temp_dir.path().join("mutants.out/log/one.log").is_file());
+        assert!(temp_dir
+            .path()
+            .join("mutants.out.old/log/two.log")
+            .is_file());
+        assert!(!temp_dir
+            .path()
+            .join("mutants.out.old/log/one.log")
+            .is_file());
     }
 }